Add changes and new files

+++ src/main/java/drinkshop/domain/Stoc.java
@@ -40,7 +40,7 @@ public class Stoc {
         this.cantitate = cantitate;
     }
 
-    public void setStocMinim(int stocMinim) {
+    public void setStocMinim(double stocMinim) {
         this.stocMinim = stocMinim;
     }
